requirements.txt
annotated-types==0.7.0
anyio==4.6.0
biopython==1.84
click==8.1.7
h11==0.14.0
idna==3.10
numpy==2.1.1
peptides==0.3.4
pydantic==2.9.2
pydantic_core==2.23.4
setuptools==75.1.0
sniffio==1.3.1
starlette==0.38.6
typing_extensions==4.12.2
wheel==0.44.0
